+++ networkit/cpp/sparsification/ChanceCorrectedTriangleScore.cpp
@@ -21,7 +21,7 @@ void ChanceCorrectedTriangleScore::run() {
 
     G->parallelForEdges([&](node u, node v, edgeid eid) {
         if ((*triangles)[eid] > 0) {
-            scoreData[eid] = (*triangles)[eid] * (G->numberOfNodes() - 2) * 1.0 / ((G->degree(u) - 1) * (G->degree(v) - 1));
+            scoreData[eid] = static_cast<double>((*triangles)[eid] * (G->numberOfNodes() - 2)) * 1.0 / static_cast<double>((G->degree(u) - 1) * (G->degree(v) - 1));
         } else if (G->degree(u) == 1 || G->degree(v) == 1) {
             scoreData[eid] = 1;
         }
